Introduce note space parameter for slide export
One can now specify whether there should be symbols for owner and duration of a wish such that the person leading the plenum can write one the slides if some is interested in becoming owner of the AK or not. The default link on the status page hides those symbols.