Introduce alternative way to input availabilities and slots
Add start and end inputs below the calendar widget that can be used to add new entries to the calendar.
Inputs are based on html5 widgets, will fallback to simple text inputs in very old browsers. Modern browsers also provide an integrated validation of the time ranges. In older browsers, this will be done latest on server side.
This implements #124