Only show necessary timeframes on AK wall
Hides hours no upcoming AK is in.
For technical reasons, only one visible area can be displayed for all days, hence hours for one day might be visible even though the are not needed (but just for another day).
For improved visibility, cutoff will be with one hour offset.
This implements #206