Introduce a graphical interface for selecting type filtering options of plan views
This displays a filter form on top of the current and next AKs view that will show both the current state of the filter and allow to change it. Changing it triggers a reload of the page with the newly selected filter options. This functionality requires javascript (as the plan does, too).